requirements.txt
# Backend
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9

# AI
openai>=1.30.0

# Detection (teammate's scope)
ultralytics>=8.2.0
opencv-python>=4.9.0
mediapipe>=0.10.0

# Communication (teammates' scope)
python-telegram-bot>=21.0
networkx>=3.2
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0
